# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 1.0.0
  description: Ratinam Crackers Business Management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /v1/returns:
    get:
      operationId: listReturns
      tags: [returns]
      summary: List returns
      security:
        - bearerAuth: []
      parameters:
        - name: type
          in: query
          schema:
            type: string
        - name: dateFrom
          in: query
          schema:
            type: string
        - name: dateTo
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Returns list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReturnListResponse"
  /v1/damage:
    post:
      operationId: logDamage
      tags: [returns]
      summary: Log damaged goods
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LogDamageBody"
      responses:
        "201":
          description: Damage logged
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessMessage"
    get:
      operationId: listDamage
      tags: [returns]
      summary: List damage logs
      security:
        - bearerAuth: []
      parameters:
        - name: locationId
          in: query
          schema:
            type: string
        - name: category
          in: query
          schema:
            type: string
        - name: dateFrom
          in: query
          schema:
            type: string
        - name: dateTo
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Damage logs
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DamageListResponse"

  # ============================================================
  # BROCHURE
  # ============================================================
...
  /v1/reports/sales:
    get:
      operationId: getSalesReport
      tags: [reports]
      summary: Get sales report
      security:
        - bearerAuth: []
      parameters:
        - name: dateFrom
          in: query
          schema:
            type: string
        - name: dateTo
          in: query
          schema:
            type: string
        - name: locationId
          in: query
          schema:
            type: string
        - name: agentId
          in: query
          schema:
            type: string
        - name: channel
          in: query
          schema:
            type: string
        - name: groupBy
          in: query
          schema:
            type: string
        - name: summary
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: Sales report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SalesReportResponse"
  /v1/reports/stock:
    get:
      operationId: getStockReport
      tags: [reports]
      summary: Get stock report
      security:
        - bearerAuth: []
      parameters:
        - name: locationId
          in: query
          schema:
            type: string
        - name: category
          in: query
          schema:
            type: string
        - name: lowStockOnly
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: Stock report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StockReportResponse"
  /v1/reports/outstanding:
    get:
      operationId: getOutstandingReport
      tags: [reports]
      summary: Get outstanding/ageing report
      security:
        - bearerAuth: []
      parameters:
        - name: agentId
          in: query
          schema:
            type: string
        - name: customerType
          in: query
          schema:
            type: string
        - name: summary
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: Outstanding report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OutstandingReportResponse"
  /v1/reports/commission:
    get:
      operationId: getCommissionReport
      tags: [reports]
      summary: Get commission report
      security:
        - bearerAuth: []
      parameters:
        - name: agentId
          in: query
          schema:
            type: string
        - name: dateFrom
          in: query
          schema:
            type: string
        - name: dateTo
          in: query
          schema:
            type: string
        - name: period
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Commission report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommissionReportResponse"
  /v1/reports/daybook:
    get:
      operationId: getDaybookReport
      tags: [reports]
      summary: Get day book
      security:
        - bearerAuth: []
      parameters:
        - name: date
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Day book
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DaybookResponse"

  # ============================================================
  # NOTIFICATIONS
  # ============================================================
...
  /v1/settings/company:
    get:
      operationId: getCompanySettings
      tags: [settings]
      summary: Get company settings
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Company settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CompanySettings"
    put:
      operationId: updateCompanySettings
      tags: [settings]
      summary: Update company settings
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCompanySettingsBody"
      responses:
        "200":
          description: Settings updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CompanySettings"
  /v1/settings/pricing:
    get:
      operationId: getPricingSettings
      tags: [settings]
      summary: Get pricing settings
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Pricing settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PricingSettings"
    put:
      operationId: updatePricingSettings
      tags: [settings]
      summary: Update pricing settings
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePricingSettingsBody"
      responses:
        "200":
          description: Settings updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PricingSettings"

  # ============================================================
  # USERS
  # ============================================================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    SuccessMessage:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
        - message
...
    PaginationMeta:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        pages:
          type: integer

    # AUTH SCHEMAS
...
    StockLevel:
      type: object
      properties:
        productId:
          type: string
        productName:
          type: string
        productCode:
          type: string
        variantId:
          type: string
        variantSize:
          type: string
        locationId:
          type: string
        locationName:
          type: string
        currentQty:
          type: integer
        reorderLevel:
          type: integer
        isLow:
          type: boolean
...
    Return:
      type: object
      properties:
        id:
          type: string
        type:
          type: string
          enum: [customer, supplier, online]
        referenceId:
          type: string
        items:
          type: array
          items:
            type: object
        reason:
          type: string
        creditAmount:
          type: number
        status:
          type: string
        createdAt:
          type: string
    ReturnListResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: array
          items:
            $ref: "#/components/schemas/Return"
        meta:
          $ref: "#/components/schemas/PaginationMeta"
    LogDamageBody:
      type: object
      required: [locationId, productId, variantId, qty, category]
      properties:
        locationId:
          type: string
        productId:
          type: string
        variantId:
          type: string
        batchNo:
          type: string
        qty:
          type: integer
        category:
          type: string
          enum: [transit, storage, handling, expired, other]
        description:
          type: string
    DamageLog:
      type: object
      properties:
        id:
          type: string
        locationId:
          type: string
        productId:
          type: string
        productName:
          type: string
        variantId:
          type: string
        qty:
          type: integer
        category:
          type: string
        description:
          type: string
        createdAt:
          type: string
    DamageListResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: array
          items:
            $ref: "#/components/schemas/DamageLog"

    # BROCHURE SCHEMAS
...
    SalesReportResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            summary:
              type: object
              properties:
                totalRevenue:
                  type: number
                totalInvoices:
                  type: integer
                avgOrderValue:
                  type: number
            byChannel:
              type: array
              items:
                type: object
                properties:
                  channel:
                    type: string
                  revenue:
                    type: number
                  count:
                    type: integer
            items:
              type: array
              items:
                type: object
    StockReportResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: array
          items:
            $ref: "#/components/schemas/StockLevel"
    OutstandingReportResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            summary:
              type: object
              properties:
                total:
                  type: number
                bucket0_30:
                  type: number
                bucket31_60:
                  type: number
                bucket61_90:
                  type: number
                bucket90plus:
                  type: number
            customers:
              type: array
              items:
                type: object
    CommissionReportResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            agents:
              type: array
              items:
                type: object
                properties:
                  agentId:
                    type: string
                  agentName:
                    type: string
                  totalSales:
                    type: number
                  commission:
                    type: number
    DaybookResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            date:
              type: string
            cashIn:
              type: number
            cashOut:
              type: number
            upiIn:
              type: number
            totalSales:
              type: number
            entries:
              type: array
              items:
                type: object

    # NOTIFICATION SCHEMAS
...
    CompanySettings:
      type: object
      properties:
        companyName:
          type: string
        gstin:
          type: string
        address:
          type: string
        phone:
          type: string
        email:
          type: string
        defaultHSN:
          type: string
        invoiceStartNumber:
          type: integer
        financialYear:
          type: string
        bankDetails:
          type: string
    UpdateCompanySettingsBody:
      type: object
      properties:
        companyName:
          type: string
        gstin:
          type: string
        address:
          type: string
        phone:
          type: string
        email:
          type: string
        defaultHSN:
          type: string
        financialYear:
          type: string
        bankDetails:
          type: string
    PricingSettings:
      type: object
      properties:
        wholesaleQtyThreshold:
          type: integer
        defaultPriceListId:
          type: string
        loyaltyEarnRate:
          type: number
        loyaltyRedemptionRate:
          type: number
        taxRate:
          type: number
    UpdatePricingSettingsBody:
      type: object
      properties:
        wholesaleQtyThreshold:
          type: integer
        defaultPriceListId:
          type: string
        loyaltyEarnRate:
          type: number
        loyaltyRedemptionRate:
          type: number
        taxRate:
          type: number

    # LOCATION SCHEMAS
...